Add code to force coverage
The RandomOrphan function and the function ecdsa_signature_parse_der_lax
in pubkey.cpp were causing non-deterministic test coverage.

In the former, if a random orphan was selected the index of which is bigger
than the max. orphan index in mapOrphanTransactions, the last orphan was
returned from RandomOrphan. If the random number generated was never large
enough, this condition would not be fulfilled and the corresponding
branch wouldn't run. The proposed solution is to force one of the 50
dependant orphans to depend on the last orphan in mapOrphanTransactions
using the newly introduced function OrphanByIndex
(and passing it a large uint256), forcing this branch to always run
at least once.

In the latter, if values for ECDSA R or S (or both) had no leading
zeros, some code would not be executed. The solution was to find a
constant value for the public key and a corresponding signature that would
be comprised of R and S values with leading zeros and calling
CPubKey::Verify at the end of the test forcing this code to always run
at least once at the end even if it hadn't throughout the test.

Removed denialofservice_tests test entry from the list of non-deterministic
tests in the coverage script.